refactor(EpisodePicker): moved EpisodePicker to higher level in MetaDetails folder
